Train 21

+++ test_util/test_installer_ccm.py
@@ -304,7 +304,7 @@ def main():
         test_util.test_runner.prepare_test_registry(
                 tunnel=test_host_tunnel,
                 test_dir=remote_dir)
-        test_util.test_runner.integration_test(
+        result = test_util.test_runner.integration_test(
                 tunnel=test_host_tunnel,
                 test_dir=remote_dir,
                 region=vpc.get_region() if vpc else DEFAULT_AWS_REGION,
@@ -323,11 +323,12 @@ def main():
     # TODO(cmaloney): add a `--healthcheck` option which runs dcos-diagnostics
     # on every host to see if they are working.
 
-    log.info("Test successsful!")
-    # Delete the cluster if all was successful to minimize potential costs.
-    # Failed clusters the hosts will continue running
-    if vpc is not None:
-        vpc.delete()
+    if result:
+        log.info("Test successsful!")
+        # Delete the cluster if all was successful to minimize potential costs.
+        # Failed clusters the hosts will continue running
+        if vpc is not None:
+            vpc.delete()
 
 
 if __name__ == "__main__":
